backend-node/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2,<3.0
scikit-learn>=1.3,<2.0
numpy>=1.24,<2.0
pandas>=2.0,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0
pillow>=10.0,<11.0
